backend/requirements.txt
Django>=5.2,<6.0
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers
psycopg2-binary
redis
channels
channels-redis
daphne
django-import-export